Redone parsing for more unambigious anonymous functions
Are we passing one or two functions to `f` in the following example?

```
f(True  -> False,
  False -> True)
```

Turns out that it was pretty ambigious. And so, I've made it slightly less
ambigious.

A single function:
```
f(True  -> False
  False -> True)
```

Two functions:
```
f(True  -> False,
  False -> True)
```

The difference being the comma after the first `False`. The change means that
clauses can't be separated by commas.