Added tagging of sum-types
As part of this change I've refactored so 'symbol' is used both for types, defs and variables.
The functionality of the tagger is quite subtle. We're parsing the syntax tree in four stages:

1. Capture all top-level types
2. Capture all non top-level types and link any mention of top level types to the types we've captured in step 1
3. Capture all top-level defs
4. Capture all variables and tag them. The function for tagging the variables checks first in
the type env if the symbol is defined. If not it uses the variable Env passed to the tag function to see if a fresh
unique variable symbol is needed or if we already have one.

This is the reason why in Step 4, we can pass an empty Env in to a pattern and still lookup types in the type env which is part of the tagging function's closure