Adding product types
I was originally going to have product types be of the form `Identifier(t1:
Type1, t2: Type2)`, this was sufficiently close to function application that
I've changed my mind and have used `Identifier: (t1: Type1, t2: Type2)`. This
form lends itself nicely to a semantic interpretation where the identifier is a
new type and the specification after the colon is the constraints of the type.

I guess this lends itself to the following alternative product types:
- `Identifier: {t1: Type1, t2: Type2}` (The product type is identified by a map)
- `Identifier: {Type1, Type2}` (The product type is identified by a set)
- `Identifier: [Type1, Type2]` (The product type is identified by a list)

I'm not sure exactly if any of these would be of any use, but the same pattern
would be very useful for defining new types:

`type temperature = Celcious: Float | Fahrenheit: Float`

In this instance, the Celcous type would be an opaque new value type which
didn't box the underlying `Float` type.