Added tagging for product types
I've fought a lot with the parser to add a convenient syntax for product and
sum types. In the end I've decided to parse product types as simple
dictionaries, while a tuple is considered a sum type. In addition for sum types
I've added the syntactic sugar of being able to specify them separated by `|`
as in `type Boolean -> True | False`.

Yecc has not been pleased. I've had to limit multi line expressions to those inside tuples, although statements mixing pipes and newlines should still be ok as in:

```
type Animals = (Cow | Cat
                Parrot | Pidgeon
                Brontosaurus)
```