This is a loooooong overdue commit of Domain logic for type checking
What's working and done'ish:
* Domain functions (diff, union, intersection)
* Domain definitions implicitly (sum, product, any, tagged, none)
* Product type definition

What's in progress:
* Scanner: Scan source code and generate domain environment
* Type generation: Generate module with functions for defined types
* Product type pattern matching
* Where clauses

There were several reasons why this was a big bite to chew. Most prominently it
turned out to be the first time I had a proper think through of the logic of
propagating domains through the source code. Out of this work came several
follow up questions:
* How is a domain defined for a function? -- After much forth and back, I've
  defined a function domain as a function of the domains of its arguments. This
  means that the domain of the function will change depending on it's inputs. It
  also means that we will scan the entire function body for every place it's
  applied.
* If a function is recursively called, how do I get the domain? -- The best
  answer I've come up with (but haven't implemented yet), is for the function
  domain to contain a set of domains of arguments it has been applied to this
  far. If the function is recursively called with the same domain, the returned
  function domain is `none`. The full function domain will be the union of its
  branches, so returning none for the recursive branches help us define the
  function type using the branches that actually returns us a type.
* How do make sure function domains are defined so any function I need the
  domain of when I scan the code has already been defined? -- We circumvent
  this problem entirely by using a function as the function domain (see the top
  question above).
* How does one define a union and intersection of the domain of a product type?
  -- This is remarkably non-trivial and I've documented my approach in the
  `type/domain.erl` source code.

These are a few I can remember, but many more have kept me up at night. I'm
hoping to be able to encapsulate some of the more subtle logic in the
`domain.erl` definitions and have the scanner work in a manner that is less
brittle to changes. I've added a bunch of unit tests to the domain functions in
the hope that they may make it easier to reason about its functionality.