Added union and intersection for {f, DomainFun} domain
This commit involves some gnarly `scanner.erl` code in order to keep `{f,
DomainFun}` as a function that takes a domain and returns a domain.

In short, to generate the environment of all the function definitions in
module, we need to pass each function an environment of all function
definitions in the module. However, because the DomainFun doesn't accept an
environment, we can't pass the environment in as we call it. Instead, we create
an awful and somewhat funny workaround:

* We generate a list of Environment functions, i.e. a list of functions that
  given the very same list of environment functions, produce an environment
  (e.g. `#{Name => {f, DomainFun}}`)
* An individual environment function when given the list of environment
  functions, just returns a `DomainFun` (i.e. a function that given argument
  domains returns the function domain), and in the closure of the domain
  function, it captures the list of environment functions
* When we call the `DomainFun`, we can easily generate the environment of all
  module functions by going through all the captured environment functions and
  pass them the list of environment functions
* Because this step happens inside the `DomainFun` and not inside the
  Environment function, we avoid triggering an infinite loop
* Instead the list of environment functions is evaluated each time we take a
  domain of type `{f, DomainFun}` and make it concrete by calling
  `DomainFun(ArgDomains)`

One consequence of this approach is that we'll need to call a function to check
domain errors from inside the scan function. The environment of a function body
isn't captured because the `DomainFun` only returns the domain.