First steps towards type generation (still not hooked in)
Here's what's in place:

* A rough structure and a few functions for generating code.
* A passing test for a generated type that returns the domain of it's variable

On a more general note, here's where I'm heading and what I have in mind:

When we scan source code to compute the domains of its variables and functions,
we make calls to a compagnion module that defines a function for each top-level
type that defines it's domain, as well as the following functions:

* `T.domain`: Returns the domain of any top-level or non-top-level type (`Cons`
  or `Nil` are non-top-level type of `List` for example). For types that cover
  over a variable, the domain `{f, DomainFun}` is used. For types that don't,
  the respective domain is used instead.
* `T.args`: Returns the domain of the arguments of any top-level or
  non-top-level type.

The domain function is computed by first running `scanner:scan` on the type
definitions, returning an environment of types mapped to their domains. In the
typegen module, we translate this environment in to a single function that
pattern matches on the qualified type name (e.g. `[List, Cons]`). For domains
of the shape `{f, DomainFun}`, we need to represent it as the computed domain
with argument parameters in place of actual domains. An example could be a Type
which adds to parameters to a boolean type: `{f, [a,b], {sum, {True, False, a,
b}}}`.

If we further add the domains to the list of arguments `[a: Domain, b: Domain]`
during the scan, we can easily compute the args function as well based on the
returned domain environment.