Added code generation for types
A type declaration is in many ways analogous to a function declaration, except
both parameters and output are domains. To give an example `type Option a -> a
| None`, defines a type function which given a domain as the paramter `a`
returns a domain. The new domain is the possible values of `a` plus an added
`None` value.

Say we defined a boolean as: `type Boolean -> True | False`. Then we can create
an optional boolean by using `Option(Boolean)` which has the domain of `True |
False | None`.

We can compute this value by calling the `Option` function in the type module
and pass it `Boolean` as an argument.

In the compiled type module we also define a `domain` function which serves to
return the type information of inner types like `True` and `False` for the type
`Boolean`. Calling `domain(Boolean/True)` will return `'Boolean/True'` (i.e.
the domain of this type is a single value).