Added scanner support for lookups
A lookup is a convenient way to extract values from a product. In the code
below I used `t { blup }` to extract the value for the `blup` key in `t`:

```
Code = "type T -> T: {blip: (Blip | Blop)
                      blup: (Blup | Blap)}
def f a
 | (t: T) -> t { blup }

> f(T(Blip, Blup)) -> Blup
```

While implementing the domain logic for this operation I ran in to the problem
that keys in pairs are treated as a special case and are not added to the
environment. This was a behavior I had introduced in order to scan product
types without polluting the environment with the product keys. At the time I
wanted to avoid that the keys pollute the global namespace.

This introduced two problems which are both manifested in the code above:
* The variable `t` in the pattern match is bound to two different tags because
  the first t is a key in a pair and not added to the environment
* The key `blup` doesn't match the key `blup` in the type. While `blup` isn't
  added to the environment while scanning `T`, it's still tagged as `{type, _,
  [T, Blup]}`. In the function body of `f` on the other hand, blup is tagged as
  `{variable, _, f_blup_sdfgs}`.

To avoid this from happening, I've let the tagger introduce a new AST type:
`{key, Line, Value}`. This value is only put in place when tagging a dictionary
to avoid e.g. type refinements from being tagged as keys.