Support for recursive type definitions
There's a few useful data structures that can be defined using recursion, like
for example a linked list which calls itself inside the definition of `Cons`:

```
type List a -> Nil | Cons: {head: a, tail: List(a)}
```

Representing the domain for a List like this tricky because its an infinite chain:

```
List = {f, (a -> {sum, #{Nil, {tagged, Cons, {product, #{head => a, tail =>
		 {sum, #{Nil, {tagged, Cons, {product, #{head => a, tail =>
		 {sum, #{Nil, {tagged, Cons, {product, ...}}}}}}}}}}}}}})}
```

To support recursive types, we need a domain representation that allows us to
denote when a domain might call itself:

```
List = {f, (a -> {sum, #{Nil, {tagged, Cons, {product, #{head => a, tail =>
		 {recur, f}}}}}})}
```

Here, `f` is a function which returns the domain of `List(a)` when called
without arguments. This allows us to keep our representation of the domain
tight, but also introduces complexity when evaluating unions or intersections
over an infinite domain without recursing forever.

To tackle this complexity, I've used a couple of different strategies:
* Recursion Path: In the `diff` function I track the `recur` functions we've
  expanded while traversing the domain structure. If we come across a recur
  function which has already been evaluated, we stop the recursion.
* Push to the front: In the `compact` function I return a `recur` function if
  any elements of a sum or product are recursive. This allows me to evaluate
  these domains inside the returned function, essentially delaying the
  compaction to the point where it is needed, similar to lazy evaluation in say
  Haskell
* Single point of entry: In all the exported functions, a single function
  definition serves as the point of entry, calling a deriviative function. In
  the case of `diff` this is necessary to introduce a stack, but for
  `intersection` and `union` it's useful because we can call `compact` when the
  function has resolved, rather than every time it gets called recursively.
  This allows us to end the call by rolling off any layers of `recur` (e.g.
  evaluate the recur domain) before returning a value.

These changes have added quite a bit of complexity to the domain functions, but
I'm hoping that outside of this module, it won't be too much of a bother.
Finger's crossed. I'm still not a 100% sure that this is the right solution to
recursive type definitions, but I'm not sure I can think of any other way to
approach the problem.