Calling of function `F` in `{f, Name, F}` is treated consistently
`{f, Name, F}` is used to encode the domain of a function. The function `F`
returns the function's domain given the domains of its arguments. In the code
so far, I've had different standards for how `F` was called. In the type
functions generated by `typegen`, `F` would be called as `F(Arg1, Arg2)`, and
in `scanner:scan` function, it would be called as `F(Stack, [Arg1, Arg2])`.

The Stack argument was needed in the scanner when a function would call itself
to prevent it from entering an infinite loop. Before calling any domain
function `F`, it would first check if `F` had already been called with the same
arguments earlier, and if so, return `none` instead of the domain.

Ideally I want any domain function `F` to be called as `F(Arg1, Arg2)`, but it
have not been able to come up with any other pallatable way to prevent infinite
recursion, so after some thought it looked to me like it came down to a choice
between:

 1) Always passing the stack as the first argument to any domain function
    (whether it needed this argument or not). This feels iffy because it would
    leak the Stack out of the scanner and complicate the interfaces of domain
    functions like `union` and `intersection`.
 2) Create distinct function domains depending on whether a stack was needed as
    an argument or not. This feels iffy for the same reasons as Option 1, but
    would add more domain types which would complicate the domain functions
    further.
 3) Passing the stack for the domain functions that needed it as part of the
    clojure of a surrounding function (i.e. a hack). This feels iffy because it
    was a hack that would make it harder to reason about code in the `scanner`
    module.

In the end I went with Option 3 because it at least compartmentalises the
increased complexity inside the `scanner` module, instead of bleeding it out to
surrounding modules.

Option 3 is implemented by wrapping `{f, Name, F}` inside the `StackFun` of
`{f, Name, StackFun}`. The `StackFun` accepts a stack as the parameter which is
then available as part of the function clojure where we define `F`.

All options involve translating a function which accepts an array of arguments
into a function which accepts the elements of the array as individual
arguments. Erlang makes this conversion a bit awkward since there aren't any
in-built options for converting a function in this way. I've kludged my way
around this limitation by creating a clause list mapping arrays up to 25
elements to the corresponding function.