Introduction of type patterns
I've just lost a really informative commit message and I'm a bit sad about it. I'm going to try to pin some of the information down again.

What's new:

* Parser support for type patterns
* Tagger support for type patterns
* Type code generation support for type patterns that consist of a type value
* (e.g. 'T')

What's happening next:

Domain Representation
---------------------

+---------+---------------+------------------------------+----------------------+
| Type    | Domain ex     | Domain Repr                  | Erlang Core Repr     |
+---------+---------------+------------------------------+----------------------+
| Value   | 'T'           | 'T'                          | 'T'                  |
| Sum     | A | B | C     | {sum, set({A, B, C})}        | {sum, A, B, C}       |
| Product | {a: A, b: B}  | {product, #{a => A, b => B}} | {product, A, b}      |
| Tagged  | T: Domain     | {tagged, 'T', Domain}        | {tagged, Domain}     |
| Func    | F(Int)        | {f, 'F', fun(a) -> ...}      | fun(a) -> ...        |
| List*   | [A, B, C]     | [A, B, C]                    | [A, B, C]            |
| Range*  | 1 ... N       | {range, 'Int', 1, N}         | {range, 'Int', 1, N} |
| Lazy*   | `T            | {lazy, fun() -> T}           | {lazy, fun() -> T}   |
+---------+---------------+------------------------------+----------------------+

The last few marked with a * haven't been implemented yet.

Type Symbols
------------

Currently at the parsing levels, types and variables are indistuingishable. This has made the code simpler, but it causes a problem when looking at pattern matching the following type function:

```
type X a
 | t -> t
 | T -> T
```

In the pattern clauses, `t` refers to a variable of any domain while `T` refers to a specific domain `T`. When generating code for this function, we treat the two differently. `t` becomes a `c_var` while `T` becomes a `c_atom` (or the domain it represents).

To fix this issue I'm going to distinguish type symbols from variable symbols at the parsing level.

Domain functions in scanner
---------------------------

Domain functions map from one or more domains to a domain. This is currently the case both in the type functions generated in typegen and the normal domains of functions generated by the scanner. However the scanner will have to return the syntax tree to generate code and as part of this, we'll need to change so it doesn't return domain functions from scanning a function anymore. It might still make sense that a domain function is part of the output, but it can't be the only output.

Application of inner types
--------------------------

Take the following type function which given a dictionary type, maps the domain under the key `s` to a domain under the key `t` and tagged by `T`:

```
type ST a
 | {s} -> T {t: s}
```

If I were to call the inner type function `T`, I'd anticipate it to accept the same variables as `ST` does, e.g. `T({s: Int}) -> T: {b: Int}`. This means that when I generate this function, I need to include the pattern matching step.