Disambiguated type symbols and variables symbols
This change makes it possible to pattern match type functions where it
won't be appearant from the context whether a symbol is a type constant
or a variable like:

```
type X a
 | T -> T
 | t -> {type: t}
```

In this type function, `t` in the pattern `t -> {type: t}` is to be interpreted
as a variable and mapped to the parameter `a`, while `T` is a type
constant and defined. This function now works.

As part of this change, I've gotten rid of the `key` construct.
Dictionaries pairs used to be parsed as a `key: expr` where `key` was
distinct from a `variable`. Given that it's clear from the context if a
variable is in a dictionary, I've changed so the key component of a
dictionary is a `variable` instead.

Another component of this change is the `IsPattern` parameter to the
`tag` function in the `tagger` module. Inside a pattern match we will
include variables introduced as dictionary keys to the environment,
while outside a pattern match we won't. Unfortunately it's non-trivial
to assess if we're tagging inside a pattern match.

One way to approach this problem is to create a separate branch of
tagging functions (`tag_pattern`), but given that we're looking at only
a single deviation, this would mean duplicating most of the existing
code, which is both error prone (bugs would have to be corrected in two
places), and hard to read. Instead I chose to include a flag which is an
ugly hack. There might be a better way to go about this, but I haven't
thought of one that didn't disgust me.