Added type function pattern matching for product domains
This commit also reverses the addition of translation module in the
previous commit. I had this great idea that transforming all domains to
tuples would make matching super easy. I hadn't taken in to account that
I wanted structural matching of products, so for example, the pattern
`{a, b}` should match the product `{a: 1, b: 2, c: 3}` despite the fact
that the latter contains a key `c` not mentioned in the pattern.

With tuples this is very hard to do, because it means that the length of
the tuple I'm matching needs to accomodate any input and so needs to
match against any length, even if we're only looking for a few specific
keys.

While a tuple representation of a product is enticing because it's
compact and so for that reason presumably both more memory efficient,
it's just not going to fly, and right now efficiency isn't my highest
priority. So instead we're using a map which erlang is fortunately very
good at partially matching. It also simplifies the code because it
removes the need to translate between a code level representation and a
runtime domain representation. At least for now.