Changed set representation from ordsets to maps
Erlang set support is not great. I've used ordsets so far and now I've
spent a lot of time changing it to maps under the presumption that I'd
need it for pattern matching. But I've just realised that this is likely
not going to be the case. We don't pattern match unions anyway.

Doh.

I'll revert this commit and move on with my life...