Added application and sum domain pattern matching
Two complications:

Sum Domains
-----------

We match a sum domain by adding a clause for each possible sum value.
This can quickly add up if we're for example matching a product of sums.
To give an example, say we have the
following types:

```
type X -> Y | Z
type P -> {x: X, xx: X}
type Test t
 | P -> Matched
 | _ -> Unmatched
```

We want the pattern match of `Test` to match all combinations of `{x: Y,
xx: Y}`, `{x: Z, xx: Z}`, `{x: Y, xx: Z}` and `{x: Z, xx: Y}`, so when
matching a product we need to take in to account that any of the product
subdomains can be a sum. This means we end up creating the amount of
pattern clauses equivalent to the cardinality of the product domain
we're matching. This isn't surprising given erlang patterns map to
values, but it's a bit unwieldy.

Type Application
----------------

We can't pattern match application. At least not in the strict sense of
a type being applied to values. At first I thought that it totally made
sense to include support for type application in patterns, since the
result of an application is a domain which we can match against.
However, given that we haven't compiled the type code at the point when
we're generating erlang core, we can't call any of the type functions.

We can work around it quite easily though, by adding support for match
statements in type expressions, like so:

```
type Args -> A | B | C
type Option a -> a | None
type Test t -> Option(t).match(Args/A -> Args/B)
```

I haven't done that yet though, but it should be pretty straightforward
once general pattern matching is in place.