Groomed AST and integrated tools for traversing it
This commit introduces the integration of `ast:traverse` in the
evaluation of code. With this introduction, a few usecases came up that
were trivially implemented using the function. I've wrapped these up in
a module called `preen`, which preens the AST after being parsed.

The api of `ast:traverse` is tricky to pin down, and I suspect that I
might end up changing it again. The crux of the traverse function is the
insight that as we traverse a tree, there's two points where
transformations can be useful:
* On the way out, where each term has yet to be traversed. At this point
we can add metadata to the context, transform symbols and check for
assumptions about the structure of the AST.
* On the way back where the children in each term are already
transformed. This is the point where it's easy to compute for example a
domain, since the domain has already been computed for all children.

The introduction of `ast:traverse` also leads to a few other changes
that I'll discuss separately:

'Let' and 'Seq' terms
---------------------

Tuples are a nightmare to traverse because each line of expressions add
to the overall environment, but only assignments add to the overall
scope for later expressions in the tuple. There hasn't otherwise been
any need to keep track of the scope and the environment separately
during traversal because the environment of a pattern also happens to be
the additional scope.

This left me with either introducing a notion of returned scope which
for all expressions was empty and for patterns would be the exact same
value as the returned environment, or alternatively convert tuples into
a `seq` and `let` expressions. A `let` expression consists of a pattern,
an expression and a body. The assignments in the pattern are included in
the scope of the body and can easily be passed along when traversing the
`let` expression.  `seq` expressions on the other hand is the sequence
of two expressions, where one expression is evaluated, the result thrown
away and subsequently the second expression is evaluated.

Using `seq` and `let` makes the AST a lot simpler to reason about and
traverse, so I'm quite pleased with this transformation. Tuples are
still used as syntactic sugar over the two other expressions and are
translated by a pass of the new `preener` module.

Key term for dict keys
----------------------

Dictionary keys are an odd size. They are
syntactic sugar for a symbol or keyword type that the language currently
doesn't have literals for, except in dictionaries. For this reason I've
represented them as variables out of laziness, but that leads to
handling all sorts of unintuitive edge-cases. They have now been
replaced with a `key` term, and a transformation step after parsing the
AST that turns the relevant symbols in to `keys`.

Confirmed Dict/Lookup patterns
------------------------------

I've dithered on whether to allow the following patterns:

```
val dict = {a: 1, b: 2}
val {a} = dict // a = 1
```

The idea being that any keyword in the dictionary can easily be assigned
to a variable of the same name as the keyword. However, I've changed my
mind because I find the mix of keywords and variables a bit confusing.
Instead the above code can be written as:

```
val dict = {a: 1, b: 2}
val {a: val_a} = dict // val_a = 1
```