Reworked Tagger
Changes:

* Handles errors such as undeclared variables
* Supports `let` terms as introduced by preener
* Added in `let_type` statement in preener to support type declarations
* Handles inner function defs and type declarations
* Added a few needed changes to ast.erl as needed
* Possibly more

This is the application of the ast module to the domain of tagging. It
was kind of hard to get right, mostly because of some subtle elements in
how types are tagged. E.g. when a type is declared inside a function, we
want to check if the name is free (i.e. not taken by a top level
declaration). To do so, we need to know if the type_def we're checking
this for is a top-level declaration (in which case it has already been
added to scope and we don't throw an error) or an inner level
declaration (in which case we throw an error). To do this, I've added
`top_level` to the list of types the the ast traverser operates with.

It felt like a long haul to get this stuff to work. I guess that's
partly just rewriting old code, but it's a bit of a drag, since this
change was caused by a chain of other changes:

* I started implementing the tagger because I was implementing a generalised way to traverse the AST
* I made a generalised way to traverse the AST because I needed to
create yet another function that did
* I needed this function because I was implementing
pattern matching for type functions
* I can't remember why I was implementing pattern matching for type
functions anymore

But it's late, and hopefully tomorrow I'll remember what took my down this
path.