implementing typegen using ast:traverse
This change reimplements type generation using the generalised framework
for traversing ASTs. It's been a trying exersice, partially because of
the Erlang compiler's reluctance to include stack traces for run-time
map errors.

One problem has been stumping me for a while, but I think I might have
come up with a pleasing resolution: If a type has both free variables
and guards, how do I invoke a subtype? For example, take the following
recursive type:

```
Blip a b
 | A, B -> Blop {k1: A, k2: B}
 | B, t -> Blop {k1: t, k2: B}
```

Say I want to construct a `Blop` element in my code. How do I invoke it?
-- Before pattern matching, I would record which variables where used in
the type following the `Blop` tag, and since I didn't have patterns
where it's possible to generate new variables, they would always be a
subset of the variables for `Blip`. This made it easy to create a type
function `Blop` with a subset of `Blip`'s arguments.

However with pattern matching we can generate new free variables that
have little to do with the variables given to `Blip`. This means the
type expression of `Blop` might cover over many more variables than
there are arguments to `Blip` and that it's hard to map the relationship
between these arguments with the variables of `Blop`.

When considering this problem, I saw three solutions:

1. Make it so we can't call `Blop` on it's own if it contains variables.
New `Blop`s can only be created indirectly through creating `Blip`s.
2. Make it so that when we construct `Blop`, we copy the pattern
matching from `Blip` to create the new free variables. Then we need some
logic to map the variables of `Blop` to the arguments of `Blip` so we
can tell which of `Blip`s arguments `Blop` should be called with.
3. Make blop a type function of the free variables present in the type
that `Blop` covers over. Then use the `scanner` to figure out the
constraints on these variables and check that when `Blop` is invoked,
these constraints are upheld, so we don't end up creating malformed
`Blop`s.

I've chosen solution 3 because of a few different reasons:
 * I think it's important to be able to create `Blop`s
 * I think it's important that the arguments to the `Blop` function are
   easy to deduce
 * I think trying to map Blop's free variables to `Blip`s arguments is
   gnarly as fuck, and I don't want to complicate my life needlessly

It's possible that checking that we don't create malformed `Blop`s
isn't so very straightforward. Scanning type statements isn't something
the scanner currently supports, but extending that capability comes
with a host of advantages, predominantly around the fact that
other domain errors in types might be caught.

Next steps:
 * Fix parsing of type clauses. I've discovered that I can't declare a
   type underneath a type function with clauses currently
 * Add support for match statements in types (should be pretty
   straightforward)
 * Implement type scanner (possibly using ast:traverse as a use-case
   for whether it's possible to convert the regular scanner)
 * Extend type generation to general code generation and reuse work to
   generate code