Reworked Parser to fix parsing defs after pattern match
I ran in to an issue where the following code wouldn't parse:

```
def F a
 | A -> B

def T -> A
```

The parser had decided that once pattern matching had started, it would
continue forever, and because all of my test code only ever had pattern
matching in the last function, I didn't discover the issue before now,
which is a bit concerning.

After prodding at it for a while without luck I worked up a bunch of
opinions about the structure of the grammar and foolheartedly decided to
implement these changes in the hopes that they would solve the issue at
hand. The result of this work is a few parsing changes:

Lookups are gone
----------------
Lookups feel like a clunky addition that I might want to reintroduce
later, but currently I'm not in the mood to support them. In patterns a
lookup `T{x}` can be replaced by a pair: `T: {x}` which also reads more
like unpacking a tagged value. In expressions a lookup is harder to
replace. The values can be unpacked by assigning to a pattern and then
used.

Patterns aren't separated by commas
-----------------------------------

Before this change, patterns were separated by commas because they were
ambigious if they weren't. A lookup pattern like `T {x}` could also be
interpreted as the two patterns `T` and `{x}`. However with their
departure and a few smaller modifications (pairs have to be wrapped in
braces).

Lambdas are sometimes wrapped in braces
---------------------------------------

Before this change I had a complicated system in place to handle the
case where a function took a lambda as an argument alongside other
parameters, where the clauses of the lambda would be separated by newlines
and then to separate the lambda from the other arguments a comma would
be needed. This was hella ambigious and I'm quite pleased to have gone
away with it. To replace it we use braces around lambda clauses,
_except_ when supplying a lambda to a function application written in
dot-style notation. Instead of writing `[1,2].map((n -> n + 1))`, it's
possible to write `[1, 2].map(n -> n + 1)`.

It's not possible however to write `f(n -> n + 1)` because it makes
parsing the function parameters ambigious.

Match statements have disappeared
---------------------------------

Instead of a built in match statement syntax, we can just add the
following function to the prelude:

```
def match a f -> f(a)
```

It can then be used as so: `a.match(n -> n + 1)`. You'll notice that
this exactly replicates what the match statement does, but without
polluting the grammar with extra terms. Hurray!