Added imports in step to compile multiple files
On the path to a small first release I'd like to be able to add a
prelude as a library file and import it. To get there I've shaped up a
basic module definition as part of the AST and specifically worked on an
import module that allows for some static checking of imports.

This work is unfinished, in that imports aren't currently used when
generating code or using the type scanner. It currently supports
scanning beam modules and source definitions to check if the imported
function is defined.

The import syntax is as follows:

```
import test/Test 	 # type import
import test/blap 	 # def import
import test/_    	 # wildcard import
import test/{blap: blup} # import 'blap' as 'blup'
import test/{Test, blap} # both 'Test' and 'blap'
```

To this will currently "work" for both erlang and kind modules compiled
as beam files. I've written in support for Kind source files as well,
but I haven't written the logic for walking through a directory and
parsing sources files, so it's currently not very useful.

I'd like to support import of types of the shape `import
Months/_` where underneath `Months` there's an enum with months as type
values. This is difficult because I don't have any information of the
values that might lie underneath `Months` at the point when I'm scanning
the imports. Currently I've added these types of imports as a `rewrite`
import, with the idea that as long as `Months` is imported, I can
rewrite the AST when the types have been scanned so `August` becomes
`Month/August`. I'm thinking that instead of doing this at a later
point, I might just logic for scanning types up ahead of where I handle
imports.

Alongside imports, a source tree in a file also supports module
declarations with exports. The syntax for this is currently:

```
module test {
  blap: <unspecified> # Export the function 'blap'
  Test: <unspecified> # Export the type 'Test'
}
```

Exports are written as a dictionary litteral and the idea is value
component will specify the type, but I'm unsure of the exact syntax.

Next steps will be to put in a bit of logic to iterate through source
files, establish a map of them, parse the types of the module and then
go through the imports. I'm thinking it might be useful to distuingish
between `internal` and `external` types and variables so at the point
we're doing codegen or scan domains we can easily call the correct
function.