Wired in support for multiple files and imports
This is a criminally large commit with a stupid amount of changes.
Adding support for imports (syntax, tagging, importing from source and
beam files + good errors) alongside support for multiple files turned
out to be a complicated and intricate change.

To handle an import, the compiler traverses the AST and replace function
names and types that are part of the import with qualified versions
specifying the module. This is done in the existing tagger, by creating
an import scope containing all imported values and their replacements.
This is neat because it means we can easily error if we find any
qualified variable which hasn't been imported.

Here's some of the nuances and tricky parts of this implementation:

Wildcard Imports
----------------

A wildcard import (`import lists/_`) imports all members of a module. A
module can both be a local source file or a beam file. For beam files it
turns out to be straighforward to list members, but for local modules we
need to build up a source map containing all exports in all local
sources.

To do so the parser scans a directory and builds up a map of the
modules present and their exports before treating imports. The import
module then checks if a module is present in the source map before
attempting to import the module from beam files.

Type Imports
------------

A type import, imports the members of a type in to the local scope. If
in module `m` we define type `T` as: `type T -> B | C`, then `import
m/T/_` imports `A`, `B` and `C` in to the local scope.

We handle type imports by treating each exported type as a separate
module which exports its type members. This means that for compiled
sources, a beam file is generated for each type. For local sources we
mirror this approach by creating a source map for each type present in
the source files.

The type module for type `T` in module `m` is given the beam name
`m_T.beam`, and when a type is imported, it's treated like any other
module import.

Local Type Wildcard Import
--------------------------

The members of a local type e.g. `type T -> A | B | C` can be imported by
adding `import T/_` in the file which adds `A`, `B` and `C` to scope.
Without importing `T/_`, we would refer to these types as `T/A`, `T/B`
and `T/C`.

As with type imports, we generate a source map for each type in the
local source file (also types that aren't exported).

Qualified Types Inside Other Types
----------------------------------

Even though we might have imported all type members of a type to the
local scope, it's still necessary to reference these members in a
qualified way when using them inside another type. For example:

```
type Days -> Mon | Tues | Wed | Thur | Fri | Sat | Sun

import Days/_
type Weekend -> Days/Fri | Days/Sat | Days/Sun
```

In the type `Weekend`, we can't refer to the individual days as e.g.
`Fri` and have to add where `Fri` comes from. This is because I think it
would be confusing if a type member was sometimes created and sometimes
resused depending on imports. I prefer that all type members are 'fresh'
in that they are created for that particular type, unless they've been
delibirately qualified.

The reason this is kind of confusing to begin with is to due with the
type syntax being very permissive. In definitions, the only place we can
create new variables are inside a pattern. At the same time we can't
refer to existing variables in a pattern. This makes it unambigious
whether the code 'creates' a name or 'refers' to a name at any given
point. For types we can both create type members and refer to other
types in the type definition.

Still Undone
------------

* Codegen doesn't respect exports. All functions are currently exported
* Type modules. Currently a module of top level types is generated by
typegen, but we also need modules for each top level type exporting the
type members of the type.